requirements.txt
anthropic
httpx
python-dotenv
pyyaml
